+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * 4.19.0.pre
   * New Features:
     * `server`: add support for `async` from SDK
+    * **global**: #156 support sending folders with httpgw
   
 * 4.18.1
   * New Features: